+++ CHANGELOG.md
@@ -24,12 +24,14 @@
 ### Changes
 - Support for disabling operation saving based on disk utilization ([#1496](https://github.com/neptune-ai/neptune-client/pull/1496))
 
+### Fixes
+- Cast unsupported floats(infinity, negative infinity, nan) in `stringify_unsupported()` ([#1509](https://github.com/neptune-ai/neptune-client/pull/1509))
+
 
 ## neptune 1.8.1
 
 ### Fixes
 - Fixed SSL-related error handling ([#1490](https://github.com/neptune-ai/neptune-client/pull/1490))
-
 
 ## neptune 1.8.0